requirements.txt
Django == 3.1
psycopg2-binary == 2.8.6
wheel
Pillow == 8.0.1
djangorestframework==3.12
django-filter
djangorestframework_simplejwt
djangorestframework-jsonapi == 4.0.0

